Added vector math library

+++ include/drivesystems/sensor.hpp
@@ -1,6 +1,9 @@
 #include "main.h"
+#include <vector>
 
 extern pros::Imu inertial_sensor;
+// extern std::vector<pros::c::imu_gyro_s_t> rotation_list;
 extern std::vector<double> rotation_list;
 extern std::vector<pros::c::imu_accel_s_t> acceleration_list;
-extern std::vector<double> timestamp_list;
+extern std::vector<uint64_t> timestamp_list;
+
